requirements/base.txt
psycopg2-binary==2.9.1
django==3.1
django-model-utils==4.0.0